Client (#16)
* step-by-step debugging

* client trait

* Machine client and runtime

* c64 client

* runtime for client

+++ crates/machine/src/machine/machine_config.rs
@@ -2,7 +2,7 @@ pub struct MachineConfig {
     pub ram_size: usize,
     pub rom_size: usize,
     pub max_time: Option<u64>,
-    pub max_cycles: Option<u64>,
+    pub max_cycles: Option<u128>,
     pub exit_on_addr: Option<u16>,
     pub exit_on_brk: bool,
     pub disassemble: bool,
